rbac

+++ backend/controllers/PostController.php
@@ -7,6 +7,7 @@ use common\models\Post;
 use common\models\PostSearch;
 use yii\filters\AccessControl;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
@@ -80,6 +81,9 @@ class PostController extends Controller
      */
     public function actionCreate()
     {
+    	if (!Yii::$app->user->can('CreatePost')) {
+    		throw new ForbiddenHttpException('Not Allowed');
+		}
         $model = new Post();
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['view', 'id' => $model->id]);
@@ -99,6 +103,9 @@ class PostController extends Controller
      */
     public function actionUpdate($id)
     {
+		if (!Yii::$app->user->can('UpdatePost')) {
+			throw new ForbiddenHttpException('Not Allowed');
+		}
         $model = $this->findModel($id);
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['view', 'id' => $model->id]);
@@ -118,6 +125,9 @@ class PostController extends Controller
      */
     public function actionDelete($id)
     {
+		if (!Yii::$app->user->can('DeletePost')) {
+			throw new ForbiddenHttpException('Not Allowed');
+		}
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
